pkg/ssh: rename pkg/ssh to pkg/sshutil
