+++ pkg/sshutil/sshutil.go
@@ -1,4 +1,4 @@
-package ssh
+package sshutil
 
 import (
 	"context"
